style: minor renaming and addressing PR comments
