+++ testcontainers/src/core.rs
@@ -10,6 +10,6 @@ pub(crate) mod client;
 pub(crate) mod containers;
 pub(crate) mod env;
 pub(crate) mod logs;
+pub(crate) mod macros;
 pub(crate) mod network;
 pub(crate) mod ports;
-pub(crate) mod utils;
