+++ testcontainers/src/runners/async_runner.rs
@@ -11,6 +11,21 @@ use std::collections::HashMap;
 
 #[async_trait]
 /// Helper trait to start containers asynchronously.
+///
+/// ## Example
+///
+/// ```rust
+/// use testcontainers::{core::WaitFor, runners::AsyncRunner, GenericImage};
+///
+/// #[tokio::test]
+/// async fn test_redis() {
+///     let container = GenericImage::new("redis", "7.2.4")
+///         .with_exposed_port(6379)
+///         .with_wait_for(WaitFor::message_on_stdout("Ready to accept connections"))
+///         .start()
+///         .await;
+/// }
+/// ```
 pub trait AsyncRunner<I: Image> {
     /// Starts the container and returns an instance of `ContainerAsync`.
     async fn start(self) -> ContainerAsync<I>;
